% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-commands.R
\name{apt_get_install}
\alias{apt_get_install}
\title{Generate an installation command for the system dependencies of the given packages}
\usage{
apt_get_install(packages, distribution = "ubuntu", release = NULL)
}
\arguments{
\item{packages}{Character vector of packages. Alternatively, the file path of
an \code{renv} lockfile, or directory containing \code{renv.lock}. In this case, the
CRAN packages will be extracted from the lockfile and used as input.}

\item{distribution}{Character. Operating system or distribution of the target
system. For example, "ubuntu" or "centos".}

\item{release}{Character. Release of the operating system, eg. "18.04". It's
not clear what this defaults to, but it is likely to be the latest release.}
}
\value{
A character vector of required packages
}
\description{
Generate an installation command for the system dependencies of the given packages
}
\details{
This queries the RStudio Package Manager API
(\url{https://packagemanager.rstudio.com/__api__/swagger/index.html}).
This API may change in the future.
}
\examples{
\dontrun{
apt_get_install("tidyverse", distribution = "centos")
apt_get_install(
  c("plumber", "rmarkdown"),
  distribution = "ubuntu",
  release = "20.04"
)
}
}
