% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-sysreqs.R
\name{get_sysreqs_json}
\alias{get_sysreqs_json}
\title{Query the public RStudio Package}
\usage{
get_sysreqs_json(packages, distribution = "ubuntu", release = NULL)
}
\arguments{
\item{packages}{Character vector of packages. Alternatively, the file path of
an \code{renv} lockfile, or directory containing \code{renv.lock}. In this case, the
CRAN packages will be extracted from the lockfile and used as input.}

\item{distribution}{Character. Operating system or distribution of the target
system. For example, "ubuntu" or "centos".}

\item{release}{Character. Release of the operating system, eg. "18.04". It's
not clear what this defaults to, but it is likely to be the latest release.}
}
\value{
An httr response, parsed as a JSON
}
\description{
Query the public RStudio Package
}
\examples{
\dontrun{
get_sysreqs_json("tidyverse", distribution = "centos")
get_sysreqs_json(
  c("plumber", "rmarkdown"),
  distribution = "ubuntu",
  release = "20.04"
)
}
}
\keyword{internal}
