% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-sysreqs.R
\name{interpret_as_list_of_packages}
\alias{interpret_as_list_of_packages}
\title{Interpret argument as a list of packages}
\usage{
interpret_as_list_of_packages(packages)
}
\arguments{
\item{packages}{}
}
\value{
A (possibly empty) character vector.
}
\description{
The sysreqs functions take a variety of inputs: either a vector of packages,
a file path to an \code{renv} lockfile, or a path to a directory containing
\code{renv.lock}. This function determines what the intended input is, after
validating that the input is a character vector.
}
\keyword{internal}
